make ClassId 16bit. that's really enough for now

+++ src/game/ClassInfo.h
@@ -10,7 +10,7 @@
 
 struct BaseObject;
 
-typedef uint32_t ClassId;
+typedef uint16_t ClassId;
 
 struct ClassInfo {
 	ClassInfo() : id(-1) {}
